+++ processors/config.go
@@ -6,7 +6,8 @@ import (
 	gostatic "github.com/piranha/gostatic/lib"
 )
 
-var reSeparator = regexp.MustCompile(`(?m:^----\r?\n)`)
+var shortSeparator = regexp.MustCompile(`(?m:^----?\r?\n)`)
+var oldSeparator = regexp.MustCompile(`(?m:^----\r?\n)`)
 
 type ConfigProcessor struct {
 }
@@ -28,14 +29,36 @@ func (p *ConfigProcessor) Mode() int {
 }
 
 func ProcessConfig(page *gostatic.Page, args []string) error {
-	parts := reSeparator.Split(page.Content(), 2)
-	if len(parts) != 2 {
-		// no configuration, well then...
-		page.PageHeader = *gostatic.NewPageHeader()
-		return nil
-	}
+	loc := shortSeparator.FindStringIndex(page.Content())
 
-	page.PageHeader = *gostatic.ParseHeader(parts[0])
-	page.SetContent(parts[1])
+	if (loc != nil) && (loc[0] == 0) {
+		// this branch parses Github-style frontmatter, i.e.
+		//
+		//     ---
+		//     date: 1234-12-12
+		//     ---
+		parts := shortSeparator.Split(page.Content(), 3)
+		// page starts with a separator but then no second separator? This means
+		// no configuration is present.
+		if len(parts) != 3 {
+			page.PageHeader = *gostatic.NewPageHeader()
+			return nil
+		}
+		page.PageHeader = *gostatic.ParseHeader(parts[1])
+		page.SetContent(parts[2])
+	} else {
+		// this branch parses old gostatic-style frontmatter, i.e.
+        //
+		//     date: 1234-12-12
+		//     ----
+		parts := oldSeparator.Split(page.Content(), 2)
+		if len(parts) != 2 {
+			// no separator to split content? No configuration is present then.
+			page.PageHeader = *gostatic.NewPageHeader()
+			return nil
+		}
+		page.PageHeader = *gostatic.ParseHeader(parts[0])
+		page.SetContent(parts[1])
+	}
 	return nil
 }